# Fixes #2041: Improved performance of and cleaned up equality test for CIM objects