… pywbem/cim_obj.py