This is the second stage of addressing issue #2041, in which the actual performance improvement gets implemented.

Original equality performance:
* CIMInstanceName with two keybindings, last one different: 14.077 us (stddev 5.2% in 100 runs with 14 outliers)
* CIMClass with 10 properties, last one different: 197.023 us (stddev 3.0% in 100 runs with 11 outliers)

Improved equality performance:
* CIMInstanceName with two keybindings, last one different: 12.952 us (stddev 4.6% in 100 runs with 11 outliers)
* CIMClass with 10 properties, last one different: 173.040 us (stddev 3.5% in 100 runs with 7 outliers)

As a result, this change is not a significant performance improvement, but it did clean up the code and eliminated the old comparator-based logic.

**Update** related to Karl's suggestion in a review comment to merge 4 lines of conditional into an expression:

This makes the improvement slightly better. Note that there are still variations in the measured result:

With the original code (4 lines of conditional):
* CIMInstanceName with two keybindings, last one different: 13.448 us (std.dev. 4.3% in 100 runs with 12 outliers)
* CIMClass with 10 properties, last one different: 176.183 us (std.dev. 2.8% in 100 runs with 14 outliers)

Latest code in this PR (4 lines condensed into one expression):
* CIMInstanceName with two keybindings, last one different: 13.703 us (std.dev. 6.1% in 100 runs with 12 outliers)
* CIMClass with 10 properties, last one different: 182.927 us (std.dev. 3.2% in 100 runs with 17 outliers)